separate skeleton
