+++ src/lib/zones/mod.rs
@@ -1,3 +1,4 @@
+pub mod skeleton;
 pub mod virtual_line;
 pub mod zones;
-pub use self::{virtual_line::*, zones::*, zones::geometry::*, zones::geojson::*};
+pub use self::{skeleton::*, virtual_line::*, zones::*, zones::geometry::*, zones::geojson::*};
